Fixes for newsletters export and support for recognitions and photo gallery exports
