+++ Makefile
@@ -1,17 +1,21 @@
 SHELL := /bin/bash -o pipefail
 
+PDF_EXPORTER=pdf-exporter
 NEWSLETTERS=newsletters
 NEWSLETTERS_PDF=$(NEWSLETTERS)/pdf
+RECOGNITIONS=recognitions
+PHOTO_GALLERY=photo_gallery
 
 .PHONY: test test-* format build
 
 format:
-	pushd $(NEWSLETTERS) && prettier --write download.js && popd
+	pushd $(PDF_EXPORTER) && prettier --write download.js && popd
 	gofmt -w .
 
 install-ubuntu-libs:
 	sudo add-apt-repository -y ppa:longsleep/golang-backports
 	sudo apt update -y
+	sudo apt install -y golang-go
 	curl -fsSL https://deb.nodesource.com/setup_20.x | sudo bash -
 	sudo apt-get install -y nodejs libatk1.0-0 libc6 libcairo2 \
 							libcups2 libdbus-1-3 libexpat1 libfontconfig1 libgbm1 libgcc1 \
@@ -32,8 +36,18 @@ install-tools: install-ubuntu-libs install-npm-tools
 download-newsletters:
 	rm -Rf output/$(NEWSLETTERS) reports/$(NEWSLETTERS)
 	mkdir -p output/$(NEWSLETTERS)
-	pushd $(NEWSLETTERS) && node download.js && popd
+	pushd $(PDF_EXPORTER) && node download.js && popd
 	pushd $(NEWSLETTERS_PDF) && go run main.go && popd
+
+download-recognitions:
+	rm -Rf output/$(RECOGNITIONS) reports/$(RECOGNITIONS)
+	mkdir -p output/$(RECOGNITIONS)
+	pushd $(PDF_EXPORTER) && node download.js && popd
+
+download-photo-gallery:
+	rm -Rf output/$(PHOTO_GALLERY) reports/$(PHOTO_GALLERY)
+	mkdir -p output/$(PHOTO_GALLERY)
+	pushd $(PDF_EXPORTER) && node download.js && popd
 
 test: download-newsletters
