+++ README.md
@@ -1,4 +1,4 @@
-# Toastmasters International Club DCP Status Checker
+# Export Newsletters, Recognitions and Photo Gallery from District 9 Toastmasters
 Check the Club DCP status page to generate the current DCP report in Yaml and image formats.
 
 ## Install tools
@@ -11,10 +11,23 @@ Format the code
 
 `make format`
 
-## Generate Report
+## Download Newsletters
 Download the Newsletters in pdf format
 
 `make download-newsletters`
 
-![News Letter Export Results](./reports/newsletters/export_results.csv)
+[Newsletters Export Results](./reports/newsletters/export_results.csv)
 
+## Download Recognitions
+Download the Recognitions in pdf format
+
+`make download-recognitions`
+
+[Recognitions Export Results](./reports/recognitions/export_results.csv)
+
+## Download Photo Gallery
+Download the Photo Gallery in pdf format
+
+`make download-photo-gallery`
+
+[Photo Gallery Export Results](./reports/photo-gallery/export_results.csv)
