+++ reports/recognitions/export_results.csv
@@ -1,0 +1,25 @@
+Text,File Name,Status
+Toastmasters of the Year,TOASTMASTERS_OF_THE_YEAR.pdf,success
+Presidential Citation Winners,PRESIDENTIAL_CITATION_WINNERS.pdf,success
+World Champions of Public Speaking,WORLD_CHAMPIONS_OF_PUBLIC_SPEAKING.pdf,success
+International Directors; Region Advisors from D9,INTERNATIONAL_DIRECTORS_REGION_ADVISORS_FROM_D9.pdf,success
+District International Speech Contest Winners,DISTRICT_INTERNATIONAL_SPEECH_CONTEST_WINNERS.pdf,success
+District Humorous Speech Contest Winners,DISTRICT_HUMOROUS_SPEECH_CONTEST_WINNERS.pdf,success
+District Evaluation Contest Winners,DISTRICT_EVALUATION_CONTEST_WINNERS.pdf,success
+District Table Topics Contest Winners,DISTRICT_TABLE_TOPICS_CONTEST_WINNERS.pdf,success
+Area Directors/Governors of the Year,AREA_DIRECTORSGOVERNORS_OF_THE_YEAR.pdf,success
+Division Directors/Governors of the Year,DIVISION_DIRECTORSGOVERNORS_OF_THE_YEAR.pdf,success
+Triple-Crown Winners: Most Recent Year,TRIPLE-CROWN_WINNERS_MOST_RECENT_YEAR.pdf,success
+Conference History,CONFERENCE_HISTORY.pdf,success
+Current Year Education Awards,CURRENT_YEAR_EDUCATION_AWARDS.pdf,failure
+Distinguished District History,DISTINGUISHED_DISTRICT_HISTORY.pdf,success
+Charter History Since 2015-2016 + Mentors; Sponsors,CHARTER_HISTORY_SINCE_2015-2016_MENTORS_SPONSORS.pdf,success
+Pathways Proficient Awards,PATHWAYS_PROFICIENT_AWARDS.pdf,success
++ Pathways Mentor Program Completion,_PATHWAYS_MENTOR_PROGRAM_COMPLETION.pdf,success
+Distinguished Clubs,DISTINGUISHED_CLUBS.pdf,success
+District Photo Gallery,DISTRICT_PHOTO_GALLERY.pdf,success
+Distinguished Toastmasters (DTMs),DISTINGUISHED_TOASTMASTERS_DTMS.pdf,success
+Club Charter Dates,CLUB_CHARTER_DATES.pdf,success
+Special Awards,SPECIAL_AWARDS.pdf,success
+Membership Campaigns + Member Sponsors,MEMBERSHIP_CAMPAIGNS_MEMBER_SPONSORS.pdf,success
+Past District Governors and Directors,PAST_DISTRICT_GOVERNORS_AND_DIRECTORS.pdf,success
